requirements.txt
gpiod==2.2.1
greenlet==3.1.1
pillow==11.0.0
playwright==1.48.0
pyee==12.0.0
spidev==3.6
typing_extensions==4.12.2
